# static-ui/demo/demo/openapi.yaml
openapi: 3.0.3
info:
  title: Static UI Bug Tracker API
  version: 0.1.0
  description: >
    Simplified API contract for backend implementation.
    Covers auth, users, workspaces, bugs, comments, attachments, dashboard.
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users (ADMIN)
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      tags: [Users]
      summary: Create user (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserReq'
      responses:
        '200':
          description: Created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Validation or duplicate email
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: User not found
    put:
      tags: [Users]
      summary: Update user (ADMIN)
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserReq'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: User not found
    delete:
      tags: [Users]
      summary: Delete user (ADMIN)
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Delete success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Delete success
        '409':
          description: User is referenced by bugs/comments
  /users/upload:
    post:
      tags: [Users]
      summary: Upload file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                fileData:
                  type: string
                  format: binary
      responses:
        '200':
          description: Relative file URL
          content:
            text/plain:
              schema:
                type: string
                example: /users/file/avatar.png
  /users/file/{filename}:
    get:
      tags: [Users]
      summary: Download uploaded file
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
  /api/workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces
      responses:
        '200':
          description: Workspace list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'
    post:
      tags: [Workspaces]
      summary: Create workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceReq'
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
  /api/workspaces/{id}:
    get:
      tags: [Workspaces]
      summary: Get workspace detail
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Workspace detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          description: Workspace not found
    put:
      tags: [Workspaces]
      summary: Update workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceReq'
      responses:
        '200':
          description: Workspace updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
    delete:
      tags: [Workspaces]
      summary: Delete workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '204':
          description: Workspace deleted
  /api/workspaces/{id}/members:
    get:
      tags: [Workspaces]
      summary: List workspace members
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
    post:
      tags: [Workspaces]
      summary: Add member to workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, role]
              properties:
                userId:
                  type: integer
                role:
                  type: string
                  example: developer
      responses:
        '200':
          description: Member added
  /api/workspaces/{id}/members/{userId}:
    delete:
      tags: [Workspaces]
      summary: Remove member from workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Member removed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    WorkspaceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    CreateUserReq:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Nguyen Van A
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 4
        role:
          type: string
          example: developer
        phone:
          type: string
          example: "0912345678"
    UpdateUserReq:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 4
        role:
          type: string
        phone:
          type: string
        avatar:
          type: string
    UserDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        phone:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
...
    Workspace:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateWorkspaceReq:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateWorkspaceReq:
      type: object
      required: [name]
      properties:
        name:
          type: string
...